make conflicting routes a fatal error again
It is not sufficient to throw a warning, since developers might not
examine the logs. Removing all conflicting routes is easy for
duplicates, but requires more tracking for triplicates, however
unlikely that might be. In any case, any 'solution' feels like a
band-aid and caters too much to lazy developers.