add greins.app.GreinsLogger
This change makes it easier to use the gunicorn log for application
messages. Everything is overridable and configurable as before, but by
default greins will use a new logging class, `greins.GreinsLogger`.

The `GreinsLogger` class is just like the logger from gunicorn, except
it configures a handler on the root logger which forwards to the
handler for the 'gunicorn.error' logger and uses a formatter that
includes the package name of the source of each log message.

Reported by Mark Doliner (#14).